Added auto refresh after adding VM group
